Fixed codestyle

+++ Assets/Scripts/HealthView.cs
@@ -2,5 +2,5 @@ using UnityEngine;
 
 public abstract class HealthView : MonoBehaviour
 {
-    public abstract void Show(int current, int max);
+    public abstract void Show(float current, float max);
 }
